+++ pymatgen/analysis/tests/test_phase_diagram.py
@@ -1,6 +1,7 @@
 # Copyright (c) Pymatgen Development Team.
 # Distributed under the terms of the MIT License.
 
+import collections
 import os
 import unittest
 import warnings
@@ -165,6 +166,12 @@ class PhaseDiagramTest(unittest.TestCase):
         )
         self.assertRaises(ValueError, PhaseDiagram, entries)
 
+    def test_repr(self):
+        assert (
+            str(self.pd) == "Li-Fe-O phase diagram\n11 stable phases: \nFe, FeO, "
+            "Fe2O3, Fe3O4, LiFeO2, Li, Li2O, LiO, Li5FeO4, Li2FeO3, O"
+        )
+
     def test_dim1(self):
         # Ensure that dim 1 PDs can be generated.
         for el in ["Li", "Fe", "O2"]:
@@ -274,11 +281,11 @@ class PhaseDiagramTest(unittest.TestCase):
 
             # test raises ValueError on entry with element not in the phase diagram
             U_entry = PDEntry("U", 0)
-            with pytest.raises(ValueError, match="U1 has elements not in the phase diagram"):
+            with pytest.raises(ValueError, match="Unable to get decomposition for PDEntry : U1 with energy"):
                 method(U_entry)
 
             # test raises ValueError on entry with very negative energy
-            too_neg_entry = PDEntry("Li", -1e10)
+            too_neg_entry = PDEntry("Li", -1e6)
             match_msg = "No valid decomposition found for PDEntry : Li1 with energy"
             with pytest.raises(ValueError, match=match_msg):
                 method(too_neg_entry)
@@ -289,6 +296,26 @@ class PhaseDiagramTest(unittest.TestCase):
 
             out = method(too_neg_entry, on_error="ignore")
             assert out == expected
+
+    def test_downstream_methods_can_also_ignore_errors(self):
+        # test that downstream methods get_e_above_hull() and get_phase_separation_energy()
+        # can also ignore errors
+        too_neg_entry = PDEntry("Li", -1e6)
+        exotic_entry = PDEntry("U W", -1e6)
+
+        # get_e_above_hull
+        with pytest.raises(ValueError, match="No valid decomposition found for PDEntry "):
+            self.pd.get_e_above_hull(too_neg_entry, on_error="raise")
+        assert self.pd.get_e_above_hull(too_neg_entry, on_error="ignore") is None
+
+        with pytest.raises(ValueError, match="Unable to get decomposition for PDEntry"):
+            self.pd.get_e_above_hull(exotic_entry, on_error="raise")
+        assert self.pd.get_e_above_hull(exotic_entry, on_error="ignore") is None
+
+        # get_phase_separation_energy
+        with pytest.raises(ValueError, match="Unable to get decomposition for PDEntry"):
+            self.pd.get_phase_separation_energy(exotic_entry, on_error="raise")
+        assert self.pd.get_phase_separation_energy(exotic_entry, on_error="ignore") is None
 
     def test_get_equilibrium_reaction_energy(self):
         for entry in self.pd.stable_entries:
@@ -339,7 +366,7 @@ class PhaseDiagramTest(unittest.TestCase):
                     ),
                 )
 
-        # Test that we get correct behaviour with a polymorph
+        # Test that we get correct behavior with a polymorph
         toy_entries = {
             "Li": 0.0,
             "Li2O": -5,
@@ -617,13 +644,13 @@ class PhaseDiagramTest(unittest.TestCase):
         # test round-trip for other entry types such as ComputedEntry
         entry = ComputedEntry("H", 0.0, 0.0, entry_id="test")
         pd = PhaseDiagram([entry])
-        d = pd.as_dict()
-        pd_roundtrip = PhaseDiagram.from_dict(d)
+        pd_dict = pd.as_dict()
+        pd_roundtrip = PhaseDiagram.from_dict(pd_dict)
         self.assertEqual(pd.all_entries[0].entry_id, pd_roundtrip.all_entries[0].entry_id)
         dd = self.pd.as_dict()
         new_pd = PhaseDiagram.from_dict(dd)
-        new_dd = new_pd.as_dict()
-        self.assertEqual(new_dd, dd)
+        new_pd_dict = new_pd.as_dict()
+        self.assertEqual(new_pd_dict, dd)
         self.assertIsInstance(pd.to_json(), str)
 
     def test_read_json(self):
@@ -749,6 +776,10 @@ class PatchedPhaseDiagramTest(unittest.TestCase):
     def test_dimensionality(self):
         assert self.pd.dim == self.ppd.dim
 
+        # test dims of sub PDs
+        dim_counts = collections.Counter(pd.dim for pd in self.ppd.pds.values())
+        assert dim_counts == {3: 7, 2: 6, 4: 2}
+
     def test_get_hull_energy(self):
         for comp in self.novel_comps:
             e_hull_pd = self.pd.get_hull_energy(comp)
@@ -765,7 +796,7 @@ class PatchedPhaseDiagramTest(unittest.TestCase):
     def test_repr(self):
         assert repr(self.ppd) == str(self.ppd) == "PatchedPhaseDiagram covering 15 sub-spaces"
 
-    def test_as_from_dict(self):
+    def test_to_from_dict(self):
         ppd_dict = self.ppd.as_dict()
         assert ppd_dict["@module"] == self.ppd.__class__.__module__
         assert ppd_dict["@class"] == self.ppd.__class__.__name__
@@ -789,6 +820,37 @@ class PatchedPhaseDiagramTest(unittest.TestCase):
         entry = PDEntry("FeO", -1.23)
         with pytest.raises(ValueError, match=r"Missing terminal entries for elements \['Fe', 'O'\]"):
             PatchedPhaseDiagram(entries=[entry])
+
+    def test_contains(self):
+        for space in self.ppd.spaces:
+            assert space in self.ppd
+        unlikely_chem_space = frozenset(map(Element, "HBCNOFPS"))
+        assert unlikely_chem_space not in self.ppd
+
+    def test_getitem(self):
+        chem_space = self.ppd.spaces[0]
+        pd = self.ppd[chem_space]
+        assert isinstance(pd, PhaseDiagram)
+        assert chem_space in pd._qhull_spaces
+        assert str(pd) == "V-C phase diagram\n4 stable phases: \nC, V, V6C5, V2C"
+
+        with pytest.raises(KeyError):
+            self.ppd[frozenset(map(Element, "HBCNOFPS"))]
+
+    def test_iter(self):
+        for pd in self.ppd:
+            assert isinstance(pd, PhaseDiagram)
+        assert len(self.ppd) == len(self.ppd.pds)
+
+    def test_len(self):
+        assert len(self.ppd) == len(self.ppd.pds)
+
+    def test_setitem_and_delitem(self):
+        unlikely_chem_space = frozenset(map(Element, "HBCNOFPS"))
+        self.ppd[unlikely_chem_space] = self.pd
+        assert unlikely_chem_space in self.ppd
+        assert self.ppd[unlikely_chem_space] == self.pd
+        del self.ppd[unlikely_chem_space]  # test __delitem__() and restore original state
 
 
 class ReactionDiagramTest(unittest.TestCase):